Changed argument type of fcntl syscall from uint32 to size_t.
Since this argument may be used to pass pointers, uint32 is not
correct for 64-bit. Effectively no change on 32-bit targets, both
size_t and uint32 are unsigned long there.

+++ headers/private/kernel/vfs.h
@@ -176,7 +176,7 @@ int			_user_open_dir_entry_ref(dev_t device, ino_t inode,
 				const char *name);
 int			_user_open_dir(int fd, const char *path);
 int			_user_open_parent_dir(int fd, char *name, size_t nameLength);
-status_t	_user_fcntl(int fd, int op, uint32 argument);
+status_t	_user_fcntl(int fd, int op, size_t argument);
 status_t	_user_fsync(int fd);
 status_t	_user_flock(int fd, int op);
 status_t	_user_read_stat(int fd, const char *path, bool traverseLink,
